fix: rename mock/quick_p2p to mock/quic_p2p
